requirements.txt
# Core dependencies
langchain>=0.0.267
unstructured>=0.10.30
pymupdf>=1.22.5
pdfminer.six>=20221105
beautifulsoup4>=4.12.2
farm-haystack>=1.20.0
python-dotenv>=1.0.0

# Vector stores 
faiss-cpu>=1.7.4
chromadb>=0.4.6
astrapy>=0.5.1
cassandra-driver>=3.28.0

# Large language model interfaces
openai>=1.3.0
huggingface-hub>=0.16.4
google-generativeai>=0.3.0  # For Gemini embeddings

# Web framework for API
fastapi>=0.100.0
uvicorn>=0.23.2

# Authentication and authorization
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
permit-sdk>=0.2.0  # Permit.io SDK for authorization

# Testing
pytest>=7.4.0